Satisfy Lukas's preferences

+++ konrad/surface.py
@@ -49,7 +49,7 @@ class Surface(Component, metaclass=abc.ABCMeta):
         """
         self.albedo = albedo
         self.height = height
-        self['temperature'] = (('time',), np.array([float(temperature)]))
+        self['temperature'] = (('time',), np.array([temperature], dtype=float))
 
         # The surface pressure is initialized before the first iteration
         # within the RCE framework to ensure a pressure that is consistent
